@@ -1,6 +1,6 @@
 {
   "name": "parentssoundboard",
-  "version": "1.0.1",
+  "version": "1.0.3",
   "author": {
     "email": "kontakt@mokkapps.de",
     "name": "Michael Hoffmann",
@@ -23,47 +23,49 @@
     "e2e": "npm run e2e-build-debug && npm run e2e-test-debug"
   },
   "dependencies": {
-    "ajv": "^6.10.0",
-    "babel-eslint": "^10.0.1",
+    "@react-native-community/async-storage": "^1.6.1",
+    "@react-native-community/geolocation": "^2.0.2",
+    "ajv": "^6.10.2",
+    "babel-eslint": "^10.0.3",
     "prop-types": "^15.7.2",
-    "react": "16.8.3",
+    "react": "16.8.6",
     "react-art": "16.8.3",
     "react-dom": "16.8.3",
-    "react-native": "^0.59.1",
+    "react-native": "0.60.5",
     "react-native-easy-toast": "^1.2.0",
     "react-native-elements": "^1.1.0",
     "react-native-email": "^1.0.2",
-    "react-native-firebase": "^5.2.3",
-    "react-native-gesture-handler": "^1.1.0",
+    "react-native-firebase": "^5.5.6",
+    "react-native-gesture-handler": "^1.4.1",
     "react-native-i18n": "^2.0.15",
-    "react-native-tts": "^2.0.0",
-    "react-native-vector-icons": "^6.4.2",
-    "react-native-web": "^0.11.1",
-    "react-navigation": "^3.5.1",
-    "react-navigation-header-buttons": "^2.1.2",
-    "react-redux": "^6.0.1",
-    "redux": "^4.0.1",
-    "styled-components": "^4.1.3"
+    "react-native-tts": "^3.0.0",
+    "react-native-vector-icons": "^6.6.0",
+    "react-native-web": "^0.11.7",
+    "react-navigation": "^3.12.0",
+    "react-navigation-header-buttons": "^3.0.2",
+    "react-redux": "^7.1.1",
+    "redux": "^4.0.4",
+    "styled-components": "^4.3.2"
   },
   "devDependencies": {
     "@babel/core": "^7.0.0-beta.47",
     "babel-core": "^7.0.0-beta.47",
-    "babel-jest": "^24.5.0",
-    "babel-plugin-jest-hoist": "^24.3.0",
+    "babel-jest": "^24.9.0",
+    "babel-plugin-jest-hoist": "^24.9.0",
     "babel-plugin-module-resolver": "^3.2.0",
     "babel-preset-env": "^1.7.0",
-    "babel-preset-jest": "^24.3.0",
+    "babel-preset-jest": "^24.9.0",
     "babel-preset-react-native": "^5.0.2",
-    "detox": "^11.0.2",
-    "eslint": "^5.15.3",
-    "eslint-config-airbnb": "^17.1.0",
+    "detox": "^14.0.3",
+    "eslint": "^6.2.2",
+    "eslint-config-airbnb": "^18.0.1",
     "eslint-plugin-detox": "^1.0.0",
-    "eslint-plugin-import": "^2.16.0",
-    "eslint-plugin-jsx-a11y": "^6.2.1",
-    "eslint-plugin-react": "^7.12.4",
-    "jest": "^24.5.0",
+    "eslint-plugin-import": "^2.18.2",
+    "eslint-plugin-jsx-a11y": "^6.2.3",
+    "eslint-plugin-react": "^7.14.3",
+    "jest": "^24.9.0",
     "jest-react-native": "^18.0.0",
-    "react-test-renderer": "16.8.4"
+    "react-test-renderer": "16.9.0"
   },
   "jest": {
     "preset": "react-native",
