@@ -1,4 +1,4 @@
-import { AsyncStorage } from 'react-native';
+import AsyncStorage from '@react-native-community/async-storage';
 
 const BASE_STORAGE_KEY = '@ParentSoundboardStore';
 
